requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aiomysql==0.0.20
async-timeout==3.0.1
attrs==19.3.0
cchardet==2.1.5
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
importlib-metadata==1.4.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
more-itertools==8.1.0
multidict==4.7.4
packaging==20.0
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.4.6
pytest==5.3.2
PyYAML==5.3
six==1.13.0
SQLAlchemy==1.3.12
wcwidth==0.1.8
yarl==1.4.2
zipp==1.0.0
